fix(airframe-import): write times into takeoff_time/landing_time, not departure_time/arrival_time
Airframe logbook import (and the GPS-import fuzzy match that reused its
scorer) was parsing block/flight times into departure_time/arrival_time —
the pilot-log-facing pair (engine_time_counter_*/engine_time), fed by
pilots/logbook_import.py. It should target takeoff_time/landing_time, the
airframe-log-facing pair (flight_time_counter_*/flight_time), per
models.py's Flight docstring. Left uncorrected, an airframe import merging
onto a pilot-import placeholder would have clobbered the pilot's own
departure_time/arrival_time instead of filling in the airframe-side field.

Fixing this removed the (accidental) signal that let near-match scoring
recognise a placeholder row from a pilot import, since those rows only
ever set departure_time/arrival_time and never have takeoff_time/
landing_time to compare against. Added a graduated fallback: when the
existing row has no takeoff_time/landing_time, score against its
departure_time/arrival_time instead, in confidence bands centred on the
aircraft's own flight_counter_offset (proxying the typical block-to-
airborne time gap) rather than a flat yes/no tolerance.

+++ app/aircraft/gps_import.py
@@ -662,7 +662,7 @@ def build_geojson(trackpoints: list[TrackPoint]) -> dict[str, Any]:
 def score_gps_candidates(
     fields: dict[str, Any],
     candidates: list[Any],
-    scorer: Callable[[dict[str, Any], Any], int],
+    scorer: Callable[[dict[str, Any], Any], float],
     min_score: int,
 ) -> list[Any]:
     """Rank *candidates* (Flight rows) against a parsed GPS segment's
